Corrección métodos de clase Concurso
